+++ src/default/RandomNormalLike.c
@@ -74,8 +74,8 @@ static void RandomNormalLike_operator(struct onnx_node_t * n)
 			float ty, tx;
 			for(size_t i = 0, l = y->ndata; i < l; i++)
 			{
-				ty = (float)rand() / (RAND_MAX + 1.0f);
-				tx = (float)rand() / (RAND_MAX + 1.0f);
+				ty = (float)rand() / ((float)RAND_MAX + 1.0f);
+				tx = (float)rand() / ((float)RAND_MAX + 1.0f);
 				py[i] = float16_to_float32(pdat->mean + pdat->scale * sqrtf(-2.0f * logf(tx)) * cosf(2.0f * acosf(-1.0f) * ty));
 			}
 		}
@@ -86,8 +86,8 @@ static void RandomNormalLike_operator(struct onnx_node_t * n)
 			float ty, tx;
 			for(size_t i = 0, l = y->ndata; i < l; i++)
 			{
-				ty = (float)rand() / (RAND_MAX + 1.0f);
-				tx = (float)rand() / (RAND_MAX + 1.0f);
+				ty = (float)rand() / ((float)RAND_MAX + 1.0f);
+				tx = (float)rand() / ((float)RAND_MAX + 1.0f);
 				py[i] = pdat->mean + pdat->scale * sqrtf(-2.0f * logf(tx)) * cosf(2.0f * acosf(-1.0f) * ty);
 			}
 		}
@@ -98,8 +98,8 @@ static void RandomNormalLike_operator(struct onnx_node_t * n)
 			double ty, tx;
 			for(size_t i = 0, l = y->ndata; i < l; i++)
 			{
-				ty = (double)rand() / (RAND_MAX + 1.0f);
-				tx = (double)rand() / (RAND_MAX + 1.0f);
+				ty = (double)rand() / ((float)RAND_MAX + 1.0f);
+				tx = (double)rand() / ((float)RAND_MAX + 1.0f);
 				py[i] = pdat->mean + pdat->scale * sqrt(-2.0f * log(tx)) * cos(2.0f * acos(-1.0f) * ty);
 			}
 		}
